requirements.txt
pandas
requests
selenium
webdriver-manager
python-dotenv
google-generativeai
pytest
